+++ src/main/java/com/picimako/gherkin/toolwindow/nodetype/ContentRoot.java
@@ -83,8 +83,8 @@ public class ContentRoot extends AbstractNodeType implements CategoriesHolder {
     @Override
     public String toString() {
         return getToString(
-            () -> GherkinBundle.toolWindow("statistics.module.simplified", displayName, tagCount(), gherkinFileCount()),
-            () -> GherkinBundle.toolWindow("statistics.module.detailed", displayName, tagCount(), gherkinFileCount()));
+            () -> GherkinBundle.message("gherkin.overview.toolwindow.statistics.module.simplified", displayName, tagCount(), gherkinFileCount()),
+            () -> GherkinBundle.message("gherkin.overview.toolwindow.statistics.module.detailed", displayName, tagCount(), gherkinFileCount()));
     }
 
     /**